refactor(test): group Deep Agents integration coverage
