+++ ci/cli-test-timing-hints.json
@@ -57,6 +57,14 @@
     "src/lib/state/onboard-session.test.ts": 6062,
     "src/lib/state/portable-uninstall-retirement.test.ts": 27072,
     "src/lib/tunnel/services.test.ts": 5573,
+    "test/agents/deepagents/dcode-session-supervisor.test.ts": 7644,
+    "test/agents/deepagents/dcode-wrapper-identity.test.ts": 11176,
+    "test/agents/deepagents/deepagents-code-tui-startup-check.test.ts": 28222,
+    "test/agents/deepagents/langchain-deepagents-code-config.test.ts": 9774,
+    "test/agents/deepagents/langchain-deepagents-code-direct-module-patch.test.ts": 31234,
+    "test/agents/deepagents/langchain-deepagents-code-image-credentials.test.ts": 31365,
+    "test/agents/deepagents/langchain-deepagents-code-image.test.ts": 17504,
+    "test/agents/deepagents/langchain-deepagents-code-nemotron-profile-plugin.test.ts": 7530,
     "test/agents/hermes/hermes-kanban-snapshot.test.ts": 10528,
     "test/agents/hermes/hermes-provider-foundation.test.ts": 5353,
     "test/agents/hermes/hermes-runtime-api-key.test.ts": 9872,
@@ -94,9 +102,6 @@
     "test/cli/snapshot-shields.test.ts": 14050,
     "test/cli/tunnel-command.test.ts": 16882,
     "test/credentials/credentials.test.ts": 6328,
-    "test/dcode-session-supervisor.test.ts": 7644,
-    "test/dcode-wrapper-identity.test.ts": 11176,
-    "test/deepagents-code-tui-startup-check.test.ts": 28222,
     "test/dependency-upgrade-skill-security.test.ts": 5353,
     "test/dependency-upgrade-skill.test.ts": 19850,
     "test/dns-proxy.test.ts": 5827,
@@ -135,11 +140,6 @@
     "test/installer-sandbox-build-trust.test.ts": 6681,
     "test/installer-supervisor-manifest-trust.test.ts": 6603,
     "test/internal-cli.test.ts": 16451,
-    "test/langchain-deepagents-code-config.test.ts": 9774,
-    "test/langchain-deepagents-code-direct-module-patch.test.ts": 31234,
-    "test/langchain-deepagents-code-image-credentials.test.ts": 31365,
-    "test/langchain-deepagents-code-image.test.ts": 17504,
-    "test/langchain-deepagents-code-nemotron-profile-plugin.test.ts": 7530,
     "test/layer-import-boundaries.test.ts": 5292,
     "test/list-command-helper-uses.test.ts": 7294,
     "test/local-credential-helper-session.test.ts": 5247,
