+++ test/agents/deepagents/dcode-login-profile.test.ts
@@ -7,7 +7,7 @@ import os from "node:os";
 import path from "node:path";
 import { afterEach, describe, expect, it } from "vitest";
 
-const repoRoot = path.resolve(import.meta.dirname, "..");
+const repoRoot = path.resolve(import.meta.dirname, "../../..");
 const sourcePath = path.join(
   repoRoot,
   "agents",
