+++ test/agents/deepagents/dcode-non-interactive-json.test.ts
@@ -7,7 +7,7 @@ import {
   cleanupPackageFixtures,
   createPackageFixture,
   patchFixture,
-} from "./helpers/langchain-deepagents-code-patch-fixture";
+} from "../../helpers/langchain-deepagents-code-patch-fixture";
 
 afterEach(cleanupPackageFixtures);
