+++ test/agents/deepagents/langchain-deepagents-code-managed-model-params.test.ts
@@ -9,7 +9,7 @@ import {
   cleanupPackageFixtures,
   createPackageFixture,
   patchFixture,
-} from "./helpers/langchain-deepagents-code-patch-fixture";
+} from "../../helpers/langchain-deepagents-code-patch-fixture";
 
 afterEach(cleanupPackageFixtures);
