+++ test/agents/deepagents/langchain-deepagents-code-observability.test.ts
@@ -5,7 +5,7 @@ import { spawnSync } from "node:child_process";
 import path from "node:path";
 import { describe, expect, it } from "vitest";
 
-const repoRoot = path.resolve(import.meta.dirname, "..");
+const repoRoot = path.resolve(import.meta.dirname, "../../..");
 const modulePath = path.join(
   repoRoot,
   "agents",
