+++ test/agents/deepagents/langchain-deepagents-code-provider-label.test.ts
@@ -5,13 +5,13 @@ import { execFileSync } from "node:child_process";
 import fs from "node:fs";
 import path from "node:path";
 import { afterEach, describe, expect, it } from "vitest";
-import { patchStagedDockerfile } from "../src/lib/onboard/dockerfile-patch";
+import { patchStagedDockerfile } from "../../../src/lib/onboard/dockerfile-patch";
 import {
   cleanupPackageFixtures,
   createPackageFixture,
   managedUpstreamProviderPath,
   patchFixture,
-} from "./helpers/langchain-deepagents-code-patch-fixture";
+} from "../../helpers/langchain-deepagents-code-patch-fixture";
 
 afterEach(cleanupPackageFixtures);
