+++ test/agents/deepagents/langchain-deepagents-code-reasoning-effort.test.ts
@@ -9,7 +9,7 @@ import {
   linkManagedReasoningEffort,
   patchFixture,
   writeManagedReasoningEffort,
-} from "./helpers/langchain-deepagents-code-patch-fixture";
+} from "../../helpers/langchain-deepagents-code-patch-fixture";
 
 afterEach(cleanupPackageFixtures);
